make it a component
