+++ aquawatch_mobile_app/src/components/index.js
@@ -4,3 +4,4 @@ export { default as WQIGauge } from './WQIGauge';
 export { default as EmptyGraph } from './EmptyGraph';
 export { default as DataGraph } from './DataGraph';
 export { default as DropdownComponent } from './DropdownComponent';
+export { default as LinkComp } from './LinkComp';
